Small cleanup

+++ core/reconstruction/relativeNonCentralRansacSolver.cpp
@@ -293,7 +293,7 @@ void corecvs::RelativeNonCentralRansacSolver::writeParams(double in[])
 
 void corecvs::RelativeNonCentralRansacSolver::fit(double distanceGuess)
 {
-    if (getInliersCount() < sampleSize())
+    if ((int)getInliersCount() < sampleSize())
         return;
     bestHypothesis.shift.normalise();
     bestHypothesis.shift *= distanceGuess;
